Create PillCounter python class, register with QML
This just creates the shell of a PillCounter class that will do the
business logic of counting the pills, and advertise it's findings for
the UI to use via Qt properties.

Right now this class provides pill_count and image_format properties
that are readable from QML, and an image_path property that QML should
set to the image path the user is interested in analyzing.