Add TensorFlow Lite model
This provides instructions for converting the pytorch model into a
TensorFlow Lite model that is suitable for running on an Android phone.
It can be tested on the desktop by using the `--tflite` option at
startup.

This is the first step in getting this application to run on an Android
phone.